apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary
python-dotenv
httpx
pytest
google-api-python-client
google-auth
google-auth-oauthlib
daftlistings

